exe include test
i was testing if a compiled exe script could still call new included scripts.
Conclusion: Nope.... its a failure.
To get around this, its very likely that there will be 2 scripts running to keep everything working. one would be the GUI and all the functions with it.
The other would be the execution manager that can include new trigger files. the second script will not be compiled to exe for this to work.