Trigger Manager fully functional
The trigger manager is up and running as it should, it handles triggers way better and stores them in a better readable and more efficient way.
most options still need to be developed but it that is way easier to do now as a trigger is now made into a class object and can handle that easily.